requirements.txt
gradio
Pillow
requests
python-multipart
onnxruntime
python-dotenv
isnet-general-use
rembg[gpu]  # opcional, só se quiser usar rembg
